fix(AppLayout): change in the structure of AppLayouts, optimize Layout and TopBar code
