+++ pages/_app.tsx
@@ -6,17 +6,21 @@ import "@fontsource/public-sans/400.css";
 import "@fontsource/public-sans/500.css";
 import Head from "next/head";
 
+import AppLayout from "../components/AppLayouts/AppLayout";
+
 import theme from "../theme/index";
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
     <ChakraProvider theme={theme}>
-      <Head>
-        <title>Blobscan</title>
-        <meta name="description" content="Generated by create next app" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      <Component {...pageProps} />
+      <AppLayout>
+        <Head>
+          <title>Blobscan</title>
+          <meta name="description" content="Generated by create next app" />
+          <link rel="icon" href="/favicon.ico" />
+        </Head>
+        <Component {...pageProps} />
+      </AppLayout>
     </ChakraProvider>
   );
 }
