+++ pages/testing/index.tsx
@@ -1,16 +1,13 @@
-import { Container, Box, Heading } from "@chakra-ui/react";
-
-import { PageTopBar } from "../../components/top-bar/PageTopBar";
-import PageLayout from "../../components/layouts/PageLayout";
+import { Box, Heading } from "@chakra-ui/react";
 
 const index = () => {
   return (
     <>
-      <PageLayout>
-        <Box>
-          <Heading>TESTING PAGES</Heading>
-        </Box>
-      </PageLayout>
+      {/* <PageLayout> */}
+      <Box>
+        <Heading>TESTING PAGES</Heading>
+      </Box>
+      {/* </PageLayout> */}
     </>
   );
 };
